Crate bestia_dev_text_to_speech

bestia_dev_text_to_speech

Text-to-speech CLI using Google api
version: 0.1.30 date: 2022-04-08 author: bestia.dev repository: Github

Motivation

After creating some interesting projects and tutorials for rust development and sharing them on Github, I wanted to create youtube videos for them. Sometimes it is easier to follow a video tutorial then to read kilometers of text tutorials.
The easy part is capturing the screen with OBS Studio. Everything visual is going on on the screen. It is a computer project, after all. The hard part is the spoken text.
It is impossible to just naturally talk about what I am doing while recording the screen. Maybe somebody has the talent of fast and focused speech. I don’t. Maybe because I am not a native English speaker. I am not at home as a speaker.
And my voice sound terrible and the accent and intonation is just not what english should sound like. But I can write pretty well. It takes ages and it is very slow, but I am happy with the result.
I am a programmer and for me every problem has a software solution. This is the story when you have a hammer, everything looks like a nail.

Finally I decided to use a text-to-speech approach.

I have a (very cheap) google account that I use for a VM that hosts my domain and my website. Google is very kind to offer text-to-speech api for free for 1 million characters. I hope this could be the right solution for me. It sounds very natural. It uses some magic neural network and it sounds definitely better then me.
But it is only an api. There is no prepared application or website that I can use.

What is more fun, than write an application in Rust for my problem?

Rust CLI from template

The easiest application to write in Rust is a CLI in Linux. And honestly I don’t need more. I will have a text file as input and an mp3 file as output. Perfect for a CLI application. For the first time I will develope a new Rust application entirely inside a docker container using my project https://github.com/bestia-dev/docker_rust_development. I want to have my rust development sandboxed. I am scared to allow tools and crates from unknown developers to have full access to my underlying system: Win10+WSL2 (https://github.com/bestia-dev/win10_wsl2_debian11).
In WSL2 terminal:

cd rustprojects/docker_rust_development/
sh rust_dev_pod_after_reboot.sh

In VSCode: F1, type ssh, choose Remote-SSH:Connect to Host..., choose rust_dev_pod, type the passphrase for the SSH key. We have opened a new VSCode window that works inside the Rust development container.

I will also use for the first time my template for Rust CLI development: https://github.com/bestia-dev/bestia_dev_cargo_auto_new_cli. It separates a lib and a bin, so the project could be used as a library for other projects. It has prepared automation tasks for building the project and some samples how to use tests and examples.
In VSCode terminal:

eval $(ssh-agent)
ssh-add /home/rustdevuser/.ssh/githubssh1
mkdir ~/rustprojects
cd ~/rustprojects
git clone git@github.com:bestia-dev/bestia_dev_cargo_auto_new_cli.git
mv bestia_dev_cargo_auto_new_cli bestia_dev_text_to_speech
cd bestia_dev_text_to_speech
rm -r -f .git
rm -r docs
mv src/bin/bestia_dev_cargo_auto_new_cli.rs src/bin/bestia_dev_text_to_speech.rs
code .

Template Renaming

In Cargo.toml change the name and the description and some other info.

In VSCode terminal type cargo auto release to copy the title and description to README.md and lib.rs.

Replace the title in other files. In VSCode Ctrl+Shift+H to open the replace in all files. Search for bestia_dev_cargo_auto_new_cli, replace with bestia_dev_text_to_speech. When you click on the result list item, you can see what you are actually replacing. A small icon at the right end of the item is Replace (Ctrl+Shift+1).

For now I will leave the functions that are used in examples and tests. I will modify them, when the new functions will be ready.

Github push new project

First in VSCode open Source Control (Ctrl+Shift+G), click Initialize Repository, type the message init and Ctrl+Enter. On the Github website create a new project https://github.com/new, copy the name and the description from Cargo.toml. You don’t need other files magically created. Save the new project. Github is very kind to give us some Git commands we can use. Now in VSCode terminal copy the commands from Github (second group):

git remote add origin git@github.com:bestia-dev/bestia_dev_text_to_speech.git
git branch -M main
git push -u origin main

README.md

Open README.md and delete the text you don’t need. Don’t delete the markers for the automation. You can see the markdown preview with (Ctrl+k, v). Save the changes. Commit the changes and push in VSCode terminal: cargo auto commit_and_push "readme cleaned" Commit and push often to reduce the risk of loosing your work.

google api text-to-speech

In my googlecloud account I enabled the text-to-speech api: https://console.cloud.google.com/apis/api/texttospeech.googleapis.com
Service name texttospeech.googleapis.com
I need some credentials to access the api: https://console.cloud.google.com/apis/credentials Click +Create credentials, API key, then edit, rename to API text-to-speech, restrict to only Cloud Text-to-Speech API. Use this key in your application by passing it with the ?key=API_KEY parameter for every request. Save this api key in env variable:

 export bestia_dev_text_to_speech_api_key=YOUR_API_KEY
env
echo $bestia_dev_text_to_speech_api_key
unset bestia_dev_text_to_speech_api_key

The simplest example from google:

curl -X POST \
-H "Content-Type: application/json; charset=utf-8" \
-d @sample_files/request.json \
"https://texttospeech.googleapis.com/v1/text:synthesize?key=$bestia_dev_text_to_speech_api_key"

Because of https only the domain part of the url texttospeech.googleapis.com is visible on the wire. The rest of the url /v1/text:synthesize?key=$bestia_dev_text_to_speech_api_key is encrypted. So it looks that the api-key is secure. They are encrypted on the wire (in transport) but if either end (user or server) logs the URL to a plain text file and does not sanitize credentials… now that’s a different conversation. Browsers can save the entire url in history, but I am not using a browser.

https client

Among crates curl-rust, hyper, reqwest, Isahc, Surf and ureq, I choose ureq. It is minimal.

cargo crev reviews and advisory

We live in times of danger with supply chain attacks.

It is recommended to always use cargo-crev to verify the trustworthiness of each of your dependencies.

Please, spread this info.

You can also read crev reviews quickly on the web:

https://web.crev.dev/rust-reviews/crates/

open-source and free as a beer

My open-source projects are free as a beer (MIT license).

I just love programming.

But I need also to drink. If you find my projects and tutorials helpful,please buy me a beer donating on my paypal.

You know the price of a beer in your local bar ;-) So I can drink a free beer for your health :-)

Na zdravje! Alla salute! Prost! Nazdravlje! 🍻

Modules

utils_mod

Just an example how to create and call a module in a separate file.

Functions

format_hello_phrase

format the hello phrase

format_upper

format the hello phrase with uppercase name

post_text_to_speech